% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aquametBio-package.r
\docType{package}
\name{aquametBio-package}
\alias{aquametBio-package}
\title{aquametBio: Calculate NRSA and NLA Biological Metrics & Indicators}
\description{
Calculate fish, macroinvertebrate, and zooplankton metrics used in the National Rivers and Streams Assessment (NRSA) and the National Lakes Assessment (NLA). Also assign condition class for all three assemblages as used in NRSA/NLA. Default taxa lists as used in national surveys for fish, macroinvertebrates, and zooplankton are included in the package.
}
\keyword{internal}
