% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNLA_ZoopMMI.r
\name{calcNLA_ZoopMMI}
\alias{calcNLA_ZoopMMI}
\title{Calculate the NLA zooplankton MMI}
\usage{
calcNLA_ZoopMMI(
  inMets,
  sampID = "UID",
  ecoreg = "ECOREG",
  totlnind = "TOTL_NIND"
)
}
\arguments{
\item{inMets}{A data frame containing, at minimum, the variables
specified in the arguments for sampID, ecoreg, totlnind, and the metrics
necessary for calculation of MMI by region. If values for more than
the necessary metrics are included in the input data frame, the
unnecessary  metrics will be ignored for each given site.

The necessary metrics, by aggregated bioregion, are:

 CPL: FINE_BIO, SIDID_PIND, DOM1_300_COPE_PBIO, FAM300_NAT_NTAX,
 COLLO_PBIO, OMNI_PTAX

 EHIGH: ZOCN_DEN, SMCLAD_PBIO, COPE_NAT_DEN, COARSE_NAT_PTAX, ROT_PBIO,
 OMNI300_PTAX

 PLAINS: FINE300_NAT_PBIO, SMCLAD_NAT_PIND, COPE_RATIO_300_BIO,
 FAM300_NAT_NTAX, ROT_NTAX, COPE_HERB_PDEN

 UMW: TOTL_NAT_PIND, BOSM300_NAT_PTAX, CALAN300_NAT_BIO, FINE_PTAX,
 DOM1_ROT_PBIO, COPE_HERB300_PBIO

 WMTNS: COARSE300_NAT_PBIO, LGCLAD300_NAT_PTAX, COPE300_BIO,
 ZOFN300_NTAX, PLOIMA_PTAX, COPE_OMNI_PTAX

 Descriptions of these metrics can be found in the file
 \emph{NRSA_Invertebrate_Metric_Descriptions.pdf}, included in
 the documentation for this package.}

\item{sampID}{A character vector containing the names of all
variables in inMets that specify a unique sample. If not specified,
the default is \emph{UID}}

\item{ecoreg}{A string with the name of the aggregated bioregion variable.
Valid values that correspond to regions used in NLA are
CPL, EHIGH, PLAINS, UMW, and WMTNS.}

\item{totlnind}{A string with the name of the variable with the
total individuals in each sample.}
}
\value{
A data frame containing the variables in sampID, as well as
the scored metrics, the zooplankton MMI, and the condition class for each
sites. The variable names are ABUN_PT, CLAD_PT, COPE_PT, RICH_PT,
ROTI_PT, TROP_PT, MMI_ZOOP, ZOOP_MMI_COND.
}
\description{
This is a function that calculates
the benthic MMI as used for the National Lakes
Assessment, based on inputs of the appropriate metrics.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{survey}
