% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcZoopAllMets.r
\name{calcZoopAllMets.r}
\alias{calcZoopAllMets.r}
\title{Calculate all zooplankton metrics}
\usage{
calcZoopAllMets.r(
  indata,
  inCoarse,
  inFine,
  inTaxa,
  sampID,
  is_distinct,
  ct = "COUNT",
  biomass = "BIOMASS",
  density = "DENSITY",
  is_distinct_sub = "IS_DISTINCT_300",
  ct_sub = "COUNT_300",
  biomass_sub = "BIOMASS_300",
  sub_mod = "300",
  taxa_id = "TAXA_ID",
  nonnative = "NON_NATIVE",
  genus = "GENUS",
  family = "FAMILY",
  ffg = "FFG",
  clad_size = "CLADOCERA_SIZE",
  net_size = "NET_SIZECLS_NEW"
)
}
\arguments{
\item{indata}{A data frame containing, at minimum, the variables
specified in the arguments for sampID, is_distinct, ct, biomass,
density, ct_sub, biomass_sub, taxa_id, nonnative. The dataset
should contain only data from the combined sample type (ZONW for NLA).}

\item{inCoarse}{A data frame for coarse mesh zooplankton sample data,
containing the variables
specified in the arguments for sampID, is_distinct, ct, biomass,
density, ct_sub, biomass_sub, taxa_id, nonnative. The dataset
should contain only data from the combined sample type (ZOCN for NLA)}

\item{inFine}{A data frame for fine mesh zooplankton sample data,
containing the variables
specified in the arguments for sampID, is_distinct, ct, biomass,
density, ct_sub, biomass_sub, taxa_id, nonnative. The dataset
should contain only data from the combined sample type (ZOFN for NLA)}

\item{inTaxa}{a data frame containing taxonomic information,
including the variables PHYLUM, CLASS, SUBCLASS, ORDER, SUBORDER,
FAMILY, following the same taxonomy as the default taxa list,
\emph{zoopTaxa}. There should also be autecology traits with names
that match those for the arguments \emph{ffg}, \emph{clad_size},
\emph{net_size}, \emph{genus}, and \emph{family}.}

\item{sampID}{A character vector containing the names of all
variables in indf that specify a unique sample. If not specified,
the default is \emph{UID}}

\item{is_distinct}{A string with the name of the distinctness variable,
which is assumed to have only values of 0 or 1. If not specified,
the default is \emph{IS_DISTINCT}.}

\item{ct}{A string with the name of the count variable. This
variable should be numeric or able to be converted to numeric.}

\item{biomass}{A string with the name of the biomass variable. This
variable should be numeric or able to be converted to numeric.}

\item{density}{A string with the name of the density variable. This
variable should be numeric or able to be converted to numeric.}

\item{is_distinct_sub}{A string with the name of the distinctness
variable for the combined subsample,
which is assumed to have only values of 0 or 1. If not specified,
the default is \emph{IS_DISTINCT_300}.}

\item{ct_sub}{A string with the name of the count variable for
the combined subsample.}

\item{biomass_sub}{A string with the name of the biomass
variable for the combined subsample.}

\item{sub_mod}{A string with the modifier used in the names
for the \emph{ct_sub} and \emph{biomass_sub} variables. The
default is '300' because this is the subsample size used for
individual coarse and fine mesh samples for NLA.}

\item{taxa_id}{A string with the name of the taxon ID variable
in \emph{inCts} that matches that in \emph{inTaxa}. The default
value is \emph{TAXA_ID}.}

\item{nonnative}{A string with the name of the numeric variable
indicating a non-native taxon. A value of 1 indicates that a
taxon is non-native, and 0 native. All other values will be ignored.}

\item{genus}{A string with the name of the variable containing
genus name in the \emph{intaxa} data frame. Blank or missing
genus names will be dropped.}

\item{family}{A string with the name of the variable containing
family name in the \emph{intaxa} data frame. Blank or missing
family names will be dropped.}

\item{ffg}{A string with the name of the functional feeding group
variable in \emph{inTaxa}. Values used
in calculations include FILT, HERB, OMNI, PARA, PRED, representing
filterer, herbivore, omnivore, parasite, and predator,
respectively.}

\item{clad_size}{A string with the name of the variable in
\emph{inTaxa} indicating the size class of the cladoceran
taxon. Valid values are LARGE and SMALL.}

\item{net_size}{A string with the name of the variable in
\emph{inTaxa} indicating the net size class of a taxon.
Valid values are COARSE and FINE.}
}
\description{
This function calculates all zooplankton metrics
for a given input data frame using the inputs provided. These
inputs include the names of count, biomass, and density variables,
as well as a variable indicating distinctness of each taxon.
The default value for each argument is that used for NLA.
}
