% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcZoopBaseMetrics.r
\name{calcZoopBaseMetrics}
\alias{calcZoopBaseMetrics}
\title{Calculate base zooplankton metrics for given input dataset.}
\usage{
calcZoopBaseMetrics(
  indata,
  sampID,
  is_distinct,
  ct,
  biomass,
  density = NULL,
  inTaxa,
  taxa_id = "TAXA_ID",
  ffg,
  clad_size,
  net_size,
  nativeMetrics = FALSE,
  nonnative = NULL
)
}
\arguments{
\item{indata}{A data frame containing, at minimum, the variables
specified in the arguments for sampID, dist, ct, and taxa_id,}

\item{sampID}{A character vector containing the names of all
variables in indata that specify a unique sample. If not specified,
the default is \emph{UID}}

\item{is_distinct}{A string with the name of the distinctness variable,
which is assumed to have only values of 0 or 1. If not specified,
the default is \emph{IS_DISTINCT}.}

\item{ct}{A string with the name of the count variable. This
variable should be numeric or able to be converted to numeric.}

\item{biomass}{A string with the name of the biomass variable. This
variable should be numeric or able to be converted to numeric.}

\item{density}{A string with the name of the density variable. This
variable should be numeric or able to be converted to numeric.}

\item{inTaxa}{a data frame containing taxonomic information,
including the variables PHYLUM, CLASS, SUBCLASS, ORDER, SUBORDER,
FAMILY, following the same taxonomy as the default taxa list,
\emph{zoopTaxa}. There should also be autecology traits with names
that match those for the arguments \emph{ffg}, \emph{clad_size},
and \emph{net_size}.}

\item{taxa_id}{A string with the name of the taxon ID variable
in \emph{inCts} that matches that in \emph{inTaxa}. The default
value is \emph{TAXA_ID}.}

\item{ffg}{A string with the name of the functional feeding group
variable in \emph{inTaxa}. Values used
in calculations include FILT, HERB, OMNI, PARA, PRED, representing
filterer, herbivore, omnivore, parasite, and predator,
respectively.}

\item{clad_size}{A string with the name of the variable in
\emph{inTaxa} indicating the size class of the cladoceran
taxon. Valid values are LARGE and SMALL.}

\item{net_size}{A string with the name of the variable in
\emph{inTaxa} indicating the net size class of a taxon.
Valid values are COARSE and FINE.}

\item{nativeMetrics}{A logical argument. TRUE indicates that
the subset of metrics based on native status should be
calculated. FALSE indicates that the full set of metrics
should be calculated. The default value is FALSE. If value
is TRUE, the input data should include all taxa, as well as
an indicator variable for non-native taxa (\emph{nonnative}).}

\item{nonnative}{A string with the name of the numeric variable
indicating a non-native taxon. A value of 1 indicates that a
taxon is non-native, and 0 native. All other values will be ignored.
This variable cannot share the same name as any other variable in
\emph{inTaxa}.}
}
\value{
A data frame containing the variables in sampID and
the zooplankton metrics as additional variables. The output
metrics include number of individuals, percent individuals,
biomass, percent biomass, density, and percent density for
each of several groups based on traits \emph{inTaxa}. If
\emph{nativeMetrics} = TRUE, NAT is appended to metric names.

\href{https://github.com/USEPA/aquametBio/blob/main/inst/NLA_Zooplankton_Metric_Descriptions.pdf}{NLA_Zooplankton_Metric_Descriptions.pdf}
}
\description{
This function calculates all base metrics associated
with a given input data frame using the inputs provided. These
inputs include the names of count, biomass, and density variables,
as well as a variable indicating distinctness of each taxon.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
