% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcZoopDivMetrics.r
\name{calcZoopDivMetrics}
\alias{calcZoopDivMetrics}
\title{Calculate zooplankton diversity metrics for given input dataset.}
\usage{
calcZoopDivMetrics(
  indata,
  sampID,
  is_distinct,
  ct,
  biomass = NULL,
  density = NULL,
  suffix = ""
)
}
\arguments{
\item{indata}{A data frame containing, at minimum, the variables
specified in the arguments for sampID, dist, ct, and taxa_id
and TRIBE, as well as autecology traits with names that match those
in the arguments ffg, habit, and ptv. In addition, there
should be a variable with the name in argument taxa_id that matches
with all of those in the indf data frame}

\item{sampID}{A character vector containing the names of all
variables in indata that specify a unique sample. If not specified,
the default is \emph{UID}}

\item{is_distinct}{A string with the name of the distinctness variable,
which is assumed to have only values of 0 or 1. If not specified,
the default is \emph{IS_DISTINCT}.}

\item{ct}{A string with the name of the count variable. This
variable should be numeric or able to be converted to numeric.}

\item{biomass}{A string with the name of the biomass variable. This
variable should be numeric or able to be converted to numeric.}

\item{density}{A string with the name of the density variable. This
variable should be numeric or able to be converted to numeric.}

\item{suffix}{A string to indicate the suffix that should be added
to metric names to indicate the subgroup represented in the metric.
This suffix must match the name of a variable in \emph{indata}}
}
\value{
A data frame containing the variables in sampID and
the zooplankton metrics as additional variables. If
\emph{nativeMetrics} = TRUE, NAT is appended to metric names.
}
\description{
This function calculates all diversity metrics associated
with a given input data frame using the inputs provided.
}
