% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcZoopDomMetrics.r
\name{calcZoopDomMetrics}
\alias{calcZoopDomMetrics}
\title{Calculate zooplankton dominance metrics}
\usage{
calcZoopDomMetrics(
  indata,
  sampID,
  is_distinct,
  valsIn,
  valsOut,
  taxa_id,
  subgrp = NULL
)
}
\arguments{
\item{indata}{Input data frame, containing SAMPID as variable identifying unique samples}

\item{sampID}{A character vector containing the names of all
variables in indata that specify a unique sample. If not specified,
the default is \emph{UID}.}

\item{is_distinct}{A string with the name of the distinctness variable,
which is assumed to have only values of 0 or 1. If not specified,
the default is \emph{IS_DISTINCT}.}

\item{valsIn}{A vector with the names of numeric variables to be used
in calculating dominance.}

\item{valsOut}{A vector of output suffixes to append to dominance
metric, matched with the order of valsIn variables.}

\item{taxa_id}{A string with the name of the variable that distinctly
identifies taxa in each sample.}

\item{subgrp}{A string with the name of a subgroup in \emph{indata}, with
a valid value of 1 to indicate inclusion in the subgroup.}
}
\value{
A data frame with \emph{sampID} variables and the metric containing
the % individuals in the
dominant (topN) taxa
}
\description{
This function calculates all of the dominance metrics used
for zooplankton.
}
