% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcZoopNativeMetrics.r
\name{calcZoopNativeMetrics}
\alias{calcZoopNativeMetrics}
\title{Calculate percent native metrics}
\usage{
calcZoopNativeMetrics(indata, sampID, inputNative, inputTotals)
}
\arguments{
\item{indata}{Input data frame containing variables as identified
in the arguments for \emph{sampID} and \emph{inputVars}.}

\item{sampID}{sampID A character vector containing the names of all
variables in indf that specify a unique sample. If not specified,
the default is \emph{UID}.}

\item{inputNative}{A character vector with the names of
metrics with the totals of native taxa for various
characteristics (e.g., count, density, biomass, 300-count
subsample).}

\item{inputTotals}{A character vector with the names of
metrics with the totals of all taxa. The order must correspond
to the order of metrics listed in \emph{inputNative}.}
}
\value{
A data frame containing percent native metrics
for each of the \emph{inputNative} variables as a subset
of the \emph{inputTotals}. The names of the resulting
metrics are based on the names of the inputs. Metrics
with suffixes are replaced as follows: NIND = PIND,
NTAX = PTAX, DEN = PDEN, BIO = PBIO.
}
\description{
This function calculates percent native for
each of the variables in \emph{inputVars}.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
