% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcZoopRichnessMetrics.r
\name{calcZoopRichnessMetrics}
\alias{calcZoopRichnessMetrics}
\title{Calculate richness-related metrics}
\usage{
calcZoopRichnessMetrics(
  indata,
  sampID,
  distVars,
  nonnative,
  inTaxa,
  taxa_id = "TAXA_ID",
  genus,
  family,
  prefix = ""
)
}
\arguments{
\item{indata}{Input data frame containing variables as identified
in the arguments for \emph{sampID} and \emph{inputNames}.
Large and rare taxa should be included in the input data.}

\item{sampID}{sampID A character vector containing the names of all
variables in indata that specify a unique sample. If not specified,
the default is \emph{UID}.}

\item{distVars}{A string with the name of the distinctness
variables to use for richness, which can have valid values of
0 (non-distinct) or 1 (distinct).}

\item{nonnative}{A string with the name of the numeric variable
indicating a non-native taxon. A value of 1 indicates that a
taxon is non-native, and 0 native. All other values will be ignored.}

\item{inTaxa}{A string with the name of the data frame containing
the taxonomic data.}

\item{taxa_id}{A string with the name of the taxon ID variable
in \emph{indata} that matches that in \emph{inTaxa}. The default
value is \emph{TAXA_ID}.}

\item{genus}{A string with the name of the variable containing
genus name in the \emph{intaxa} data frame. Blank or missing
genus names will be dropped.}

\item{family}{A string with the name of the variable containing
family name in the \emph{intaxa} data frame. Blank or missing
family names will be dropped.}

\item{prefix}{A vector of the same length as distVars, containing
metric name modifiers for each input variable.}
}
\value{
A data frame containing the richness metrics that include
number of taxa for full data (native only and all taxa) and
300 count subsamples (native only and all taxa), with variations
based on the lowest taxon possible, genus level, and family level.
}
\description{
This function calculates all of the variations of
taxa richness that include species-level, genus-level, and
family-level for total counts and 300-count subsamples, for
both the full dataset and native taxa only. Only distinct taxa
are included in metrics.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
