% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcZoopTotals.r
\name{calcZoopTotals}
\alias{calcZoopTotals}
\title{Calculate zooplankton totals}
\usage{
calcZoopTotals(
  indata,
  sampID,
  is_distinct,
  inputSums,
  outputSums,
  outputTaxa = NULL
)
}
\arguments{
\item{indata}{Input data frame containing variables as identified
in the arguments for \emph{sampID} and \emph{inputNames}.}

\item{sampID}{sampID A character vector containing the names of all
variables in indata that specify a unique sample. If not specified,
the default is \emph{UID}.}

\item{is_distinct}{A string with the name of the distinctness variable,
which can have valid values of 0 (non-distinct) or 1 (distinct).}

\item{inputSums}{A vector with the names of numeric variables to be
summed. If variables are not numeric, they will be converted.
This may lead to NAs being produced if any values cannot be
converted to numeric.}

\item{outputSums}{A vector with the names of numeric variables
containing total of each input variable.}

\item{outputTaxa}{A string with the name of the variable for the
number of taxa output variable.}
}
\value{
A data frame containing the totals of input variables by
the variables in \emph{sampID}.

\href{https://github.com/USEPA/aquametBio/blob/main/inst/NLA_Zooplankton_Metric_Descriptions.pdf}{NLA_Zooplankton_Metric_Descriptions.pdf}
}
\description{
This internal function calculates totals for
counts, biomass, and density from inputs. This function is
called from within various metric calculation functions using
various data subsets.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
\keyword{internal}
