% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertZoopCts_NLA.r
\name{convertZoopCts_NLA}
\alias{convertZoopCts_NLA}
\title{Convert raw zooplankton counts to density
and biomass}
\usage{
convertZoopCts_NLA(
  inCts,
  sampID,
  sampType,
  rawCt,
  biofactor,
  tow_vol,
  vol_ctd,
  conc_vol,
  lr_taxon = NULL,
  taxa_id,
  subsample = FALSE
)
}
\arguments{
\item{inCts}{A data frame containing, at minimum, the variables
specified in the arguments for sampID, sampType, rawct, taxa_id,
biofactor, tow_vol, vol_ctd, and conc_vol.}

\item{sampID}{A character vector containing the names of all
variables in indf that specify a unique sample. If not specified,
the default is \emph{UID}.}

\item{sampType}{A string specifying the name of the variable that
indicates the sample type.}

\item{rawCt}{A string specifying the name of the variable for
the raw count values for each taxon.}

\item{biofactor}{A string specifying the name of the numeric variable
for the biomass factor for each taxon, as provided by the laboratory.}

\item{tow_vol}{A string specifying the name of the numeric
variable with the tow volume in same units as vol_ctd and conc_vol.}

\item{vol_ctd}{A string specifying the name of the numeric variable
with the volume of the sample counted in the same units as tow_vol
and conc_vol.}

\item{conc_vol}{A string specifying the name of the numeric variable
with the concentrated volume of the sample in the same units as
tow_vol and conc_vol.}

\item{lr_taxon}{A string specifying the name of the variable indicating
a large/rare taxon. These do not have values for \emph{rawCt}.
This only applies to full
samples and not subsamples, since large/rare taxa should be ignored
in creating subsamples. Value should be Y or missing (NA).}

\item{taxa_id}{A string specifying the name of the variable indicating
the taxon being used.}

\item{subsample}{A logical operator to indicate whether the rawCt values
are part of a random subsample or the original sample. Valid values
are TRUE or FALSE. Default value is FALSE.}
}
\value{
A data frame containing the \emph{sampID} and
\emph{sampType}, and \emph{taxa_id} fields, plus
BIOMASS, DENSITY, and COUNT fields.
}
\description{
This function is only used to calculate
zooplankton density and biomass values from raw count
data using tow volume, concentrated volume, volume
counted, and the biomass factor. The biomass factor is
provided by the laboratory to convert counts to biomass.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
