% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metCalcFxns.r
\name{Dominance}
\alias{Dominance}
\title{Dominance metric calculation function}
\usage{
Dominance(df, topN = 1)
}
\arguments{
\item{df}{Input data frame, containing SAMPID as variable identifying unique samples}

\item{topN}{Number specifying the top number of species to include in calculation}
}
\value{
A data frame with SAMPID and the metric containing the % individuals in the
dominant (topN) taxa
}
\description{
This function calculates % dominant organisms metric
}
\keyword{internal}
