% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepZoopCombCts_NLA.r
\name{prepZoopCombCts_NLA}
\alias{prepZoopCombCts_NLA}
\title{Aggregate zooplankton count data from coarse
and fine mesh samples into ZONW sample}
\usage{
prepZoopCombCts_NLA(
  inCts,
  sampID = "UID",
  sampType = "SAMPLE_TYPE",
  typeFine = "ZOFN",
  typeCoarse = "ZOCN",
  ct = "COUNT",
  biomass = "BIOMASS",
  density = NULL,
  taxa_id = "TAXA_ID",
  lr_taxon = NULL
)
}
\arguments{
\item{inCts}{A data frame containing, at minimum, the variables
specified in the arguments for sampID, ct, biomass, and taxa_id.
Include density if the dataset includes density.}

\item{sampID}{A character vector containing the names of all
variables in indf that specify a unique sample. If not specified,
the default is \emph{UID}.}

\item{sampType}{The name of the character vector containing the
two sample types representing fine and coarse mesh zooplankton
samples. Default value is \emph{SAMPLE_TYPE}.}

\item{typeFine}{A string containing the name of the sample type
for the fine mesh zooplankton sample. Default value is \emph{ZOFN}.}

\item{typeCoarse}{A string containing the name of the sample
type for the coarse mesh zooplankton sample. Default value is
\emph{ZOCN}.}

\item{ct}{A string with the name of the count variable. If not
specified, the default is \emph{COUNT}.}

\item{biomass}{A string with the name of the biomass variable.
If not specified, the default is \emph{BIOMASS}.}

\item{density}{A string with the name of the density variable.
If not specified, the default is \emph{NULL}.}

\item{taxa_id}{A string with the name of the taxon ID variable
in \emph{inCts} that matches that in \emph{inTaxa}. The default
value is \emph{TAXA_ID}.}

\item{lr_taxon}{A string specifying the name of the variable indicating
a large/rare taxon. These do not have values for \emph{rawCt}.
This only applies to full
samples and not subsamples, since large/rare taxa should be ignored
in creating subsamples. Value should be Y or missing (NA). Default
value is NULL.}
}
\value{
A data frame containing the \emph{sampID} and \emph{sampType}
fields, plus the \emph{taxa_id},
\emph{biomass}, \emph{density}, and \emph{ct} fields.
Taxonomy is aggregated to the level
used for the NLA MMI and distinctness is recalculated on this
aggregated dataset.To obtain a similar output for different
variations on the count, biomass, and density variables.
}
\description{
This function aggregates count data from NLA
ZOCN and ZOFN samples into a new sample type (ZONW) used in
metric calculations. In order to calculate the NLA
Zooplankton MMI, both the original and 300-count subsamples
for each sample type must be calculated. For the 300-count
subsamples, only counts and biomass apply. Repeat the function for
both the full samples and the 300-count subsamples.
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
