% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarifyCounts.r
\name{rarifyCounts}
\alias{rarifyCounts}
\title{Draw a random subsample of a specified size}
\usage{
rarifyCounts(
  inCts,
  sampID = "UID",
  abund = "COUNT",
  subsize = NULL,
  seed = NULL
)
}
\arguments{
\item{inCts}{A data frame containing, at minimum, the variables
specified in the arguments for sampID and abund, and some
variable to indicate taxon, either by an ID number or name.}

\item{sampID}{A character vector containing the names of all
variables in indf that specify a unique sample. If not specified,
the default is \emph{UID}.}

\item{abund}{A string with the name of the count variable. If not
specified, the default is \emph{COUNT}. Any rows with a missing value
for the \emph{abund} variable will be dropped in the output
dataset.}

\item{subsize}{A numeric value to indicate the size of the random
subsample to be created.}

\item{seed}{Optional number to set the random seed. This allows the
process to be repeated and receive the same results, as long as
the input data are ordered the same way.}

\item{taxa_id}{A string with the name of the taxon ID or taxon
name variable
in \emph{inCts}.}
}
\value{
A data frame containing the same fields at the input
dataset, with the \emph{abund} containing the new subsample
count for each taxon.
}
\description{
This function allows the user to create a random
subsample of a specified size from the original count data.
The data should be in a long format, with each taxon on a
separate row. The user can make this process repeatable by
specifying a random seed that can be used to reproduce the
results, as long as all of the data are in the same order.
This function was adapted from a script originally
created by John Van Sickle (rarify.r, 10 June 2005).
}
\author{
Karen Blocksom \email{Blocksom.Karen@epa.gov}
}
