% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metCalcFxns.r
\name{tolindex}
\alias{tolindex}
\title{Weight tolerance value function}
\usage{
tolindex(indata, taxalist)
}
\arguments{
\item{indata}{Input data frame, containing SAMPID as variable identifying unique samples, TOTAL,
and TAXA_ID}

\item{taxalist}{A taxalist with TAXA_ID and tolerance values (TVs)
as RESULT and PARAMETER as either PTV or TOL_VAL}
}
\value{
A data frame with SAMPID and the metric WTD_TV
}
\description{
This function calculates the the WTD_TV metric as used in NARS benthic data
}
\keyword{internal}
