% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metCalcFxns.r
\name{tolindexFish}
\alias{tolindexFish}
\title{Weight tolerance value function for fish}
\usage{
tolindexFish(indata, taxalist)
}
\arguments{
\item{indata}{Input data frame, containing SAMPID as variable identifying
unique samples, TOTAL, and TAXA_ID}

\item{taxalist}{A taxalist with TAXA_ID and tolerance values (TVs)
as RESULT and PARAMETER as either PTV or TOL_VAL}
}
\value{
A data frame with SAMPID and the metric WTD_TV
}
\description{
This function calculates the fish WTD_TV metric as used in NARS.
Only taxa with TVs are included in the proportion calculations.
}
\keyword{internal}
