% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoopDominance.r
\name{zoopDominance}
\alias{zoopDominance}
\title{Dominance metric calculation function (zooplankton)}
\usage{
zoopDominance(df, sampID = "UID", topN = 1, varIn)
}
\arguments{
\item{df}{Input data frame, containing SAMPID as variable identifying unique samples}

\item{sampID}{A character vector containing the names of all
variables in indf that specify a unique sample. If not specified,
the default is \emph{UID}.}

\item{topN}{Number specifying the top number of species to include in calculation.
The default value is 1.}

\item{varIn}{A string with the name of the variable to be used in calculating
dominance.}
}
\value{
A data frame with \emph{sampID} variables and the metric containing
the % individuals in the
dominant (topN) taxa
}
\description{
This function calculates % dominant organisms metric
for zooplankton. This is done slightly differently from other
assemblages, including rounding to 1 digit instead of 2. It also
accepts different input variables, which accommodates calculating
dominance based on count, biomass, and density.
}
\keyword{internal}
