% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{zoopTaxa}
\alias{zoopTaxa}
\title{Zooplankton Taxa}
\format{
A data frame with 581 rows and the following 18 columns:
\describe{
\item{PUBLICATION_DATE}{date the dataset was published.}
\item{TAXA_ID}{the taxa ID.}
\item{TARGET_TAXON}{name of the taxon at the level used in NRSA.}
\item{PHYLUM}{taxonomic phylum name.}
\item{CLASS}{taxonomic class name.}
\item{SUBCLASS}{taxonomic subclass name.}
\item{ORDER}{taxonomic order name.}
\item{SUBORDER}{taxonomic suborder name.}
\item{FAMILY}{taxonomic family name.}
\item{SUBFAMILY}{taxonomic subfamily name.}
\item{GENUS}{taxonomic genus name.}
\item{SPECIES}{taxonomic species name.}
\item{SUBSPECIES}{taxonomic subspecies name.}
\item{FFG}{Functional feeding group trait (OMNI = Omnivore,
HERB = Herbivore, PARA = Parasite, FILT = Filterer,
PRED = Predator, UNK = Unknown)}
\item{NON_TARGET}{Non-target taxon for NLA purposes (Y or NA)}
\item{NON_NATIVE}{Non-native status. If value is ALL, non-native
in all states; if value is a list of states, states where taxon is
non-native; if missing, taxon considered native.}
\item{NET_SIZECLS_NEW}{Mesh size class assigned to taxon (COARSE/FINE)
for NLA}
\item{CLADOCERA_SIZE}{Size class of cladoceran taxon (LARGE/SMALL)}
}
}
\usage{
zoopTaxa
}
\description{
A dataset containing zooplankton taxa.
}
\note{
Unless a data frame name is provided for argument inTaxa to the
zooplankton metric functions, this dataset provides the inTaxa data frame.
}
\examples{
data(zoopTaxa)
head(zoopTaxa)
}
\keyword{datasets}
